+++ modules/ui-material/src/dynamic-form-material.component.spec.ts
@@ -14,7 +14,7 @@ import { DynamicFormMaterialComponent, DYNAMIC_FORM_UI_MATERIAL } from "./dynami
 
 describe("DynamicFormMaterialComponent test suite", () => {
 
-    let inputModel = new DynamicInputModel({id: "testInput"}),
+    let inputModel = new DynamicInputModel({id: "testInput", maxLength: 51}),
         formModel = [
             inputModel,
             new DynamicSwitchModel({id: "testSwitch"})
@@ -47,10 +47,11 @@ describe("DynamicFormMaterialComponent test suite", () => {
 
         component.controlGroup = formGroup;
         component.model = formModel[0];
+        component.showCharacterHint = false;
 
         fixture.detectChanges();
 
-        inputElement = debugElement.query(By.css(`md-input[id='${formModel[0].id}']`));
+        inputElement = debugElement.query(By.css(`input[id='${formModel[0].id}']`));
     }));
 
     it("should initialize correctly", () => {
@@ -61,6 +62,9 @@ describe("DynamicFormMaterialComponent test suite", () => {
         expect(component.controlGroup instanceof FormGroup).toBe(true);
         expect(component.model instanceof DynamicFormControlModel).toBe(true);
         expect(component.hasErrorMessaging).toBe(false);
+        expect(component.showCharacterHint).toBe(false);
+
+        expect(component.characterCount).toBe(0);
 
         expect(component.onControlValueChanges).toBeDefined();
         expect(component.onModelDisabledUpdates).toBeDefined();